made the registaer Page
